Add cover compression on backend. Rename database table book_current in book_progress. Move database column last_read to book_progress table. Move position, mark-read, mark-unread endpoints to book/:id/progress/. Remove chapter column from book_cache table. Edit chapters system, now using navigation.

+++ backend/routes/cache.js
@@ -1,26 +1,38 @@
 const response = require("../methods/response");
 const router = require("express").Router({mergeParams: true});
+const progress = require("../services/progress");
 const cache = require("../services/cache");
 const books = require("../services/books");
 const multer = require("multer");
+const sharp = require("sharp");
 const path = require("path");
 const fs = require("fs");
 
 router.put(
     "/cover",
-    multer({dest: "./covers/"}).single("cover"),
+    multer({dest: "./covers/temp"}).single("cover"),
     async (req, res, next) => {
         try {
             const {id} = req.params;
+            if (req.file.mimetype !== "image/png" && req.file.mimetype !== "image/jpeg") {
+                response.badRequest(res, "not_image", "File is not an image. mimetype=" + req.file.mimetype);
+                return;
+            }
             // remove old cover file if exists
             const results = await books.getCoverById(id);
             if (!(!results.length || !results[0].cover)) {
                 const cover = path.join(__dirname, "..", "covers", results[0].cover);
                 fs.unlinkSync(cover);
             }
+            // create compressed file
+            const buffer = fs.readFileSync(path.join(__dirname, "..", req.file.path));
+            await sharp(buffer)
+                .webp({quality: 90})
+                .toFile("./covers/" + req.file.filename);
+            // delete temp file
+            fs.unlinkSync(path.join(__dirname, "..", req.file.path));
             // add the new one in db
-            const image = req.file.filename;
-            await cache.cover(id, image);
+            await cache.cover(id, req.file.filename);
             res.send();
         } catch (err) {
             console.error("Error on PUT /books/:id/cache/cover.");
@@ -70,52 +82,10 @@ router.put(
         try {
             const {id} = req.params;
             await cache.locations(id, JSON.stringify(req.body.locations));
-            await cache.positionNull(id);
+            await progress.positionNull(id);
             res.send();
         } catch (err) {
             console.error("Error on PUT /books/:id/cache/locations.");
-            next(err);
-        }
-    }
-);
-
-router.put(
-    "/position",
-    async (req, res, next) => {
-        try {
-            const {id} = req.params;
-            await cache.position(id, req.body.position, req.body.page);
-            res.send();
-        } catch (err) {
-            console.error("Error on PUT /books/:id/cache/position.");
-            next(err);
-        }
-    }
-);
-
-router.put(
-    "/mark-read",
-    async (req, res, next) => {
-        try {
-            const {id} = req.params;
-            await cache.read(id);
-            res.send();
-        } catch (err) {
-            console.error("Error on PUT /books/:id/cache/mark-read.");
-            next(err);
-        }
-    }
-);
-
-router.put(
-    "/mark-unread",
-    async (req, res, next) => {
-        try {
-            const {id} = req.params;
-            await cache.unread(id);
-            res.send();
-        } catch (err) {
-            console.error("Error on PUT /books/:id/cache/mark-unread.");
             next(err);
         }
     }
